+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Added
 
 - `sandstone/ImageItem` public class names `horizontal` and `vertical`
-- `sandstone/Input` props `maxLength`, `minLength` and `numericInputKind` to flexible number input
+- `sandstone/Input` props `maxLength`, `minLength` and `numberInputField` to flexible number input
 - `sandstone/Input` props `invalid` and `invalidMessage` to mirror the API of `InputField`
 - `sandstone/DatePicker` function `dateToLocaleString` to create locale-aware date strings
 - `sandstone/TimePicker` function `timeToLocaleString` to create locale-aware time strings